AutoComplete - pass data through navigation props
